requirement.txt
# Data handling
pandas
scipy
numpy<2.0

# Image processing / augmentations

opencv-python<4.11
# opencv-python-headless<4.11
Pillow
albumentations

# UI
streamlit==1.31.1
streamlit-drawable-canvas==0.9.3

#hugging face
huggingface-hub

# Deep learning framework
torch
torchvision

# Dataset & evaluation tools
tqdm
scikit-learn  # metrics (F1, confusion matrix)
matplotlib     # visualization
seaborn        # confusion matrix heatmaps

# nuScenes devkit (for nuImages dataset handling)
#nuscenes-devkit

